requirements-full.txt
# Core Dependencies
python-dotenv==1.2.1
sqlite-utils==3.39

# AI/LLM
langchain==1.2.10
langchain-core==1.2.13
langchain-groq==1.1.2
langchain-openai==1.1.10
langgraph==1.0.8
langsmith==0.7.4
groq==0.37.1

# FastAPI Backend
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
python-multipart==0.0.12

# Database
psycopg2-binary==2.9.9  # PostgreSQL (for Phase 2)

# Data Processing
pandas==2.2.3
numpy==2.1.3
polars==1.12.0  # Optional: faster alternative to pandas

# ML/AI Models
scikit-learn==1.5.2
statsmodels==0.14.4  # ARIMA
prophet==1.1.6  # Facebook forecasting
mlxtend==0.23.1  # Market basket analysis
tensorflow==2.18.0  # Optional: for LSTM
joblib==1.4.2  # Model serialization

# MLOps (Optional)
mlflow==2.17.2

# Visualization
plotly==5.24.1
matplotlib==3.9.2
seaborn==0.13.2

# Streamlit Frontend
streamlit==1.39.0

# Audio (for voice features)
sounddevice==0.5.1
soundfile==0.13.1

# Utilities
requests==2.32.3
httpx==0.27.2
